Final Changes

+++ BoardConnectAPI/APIModels/RequestModels/ProjectRequestModel.cs
@@ -22,7 +22,7 @@
         /// <summary>
         /// The progress
         /// </summary>
-        public decimal? Progress { get; set; }
+        public uint? Progress { get; set; }
 
         /// <summary>
         /// The start date
