+++ BoardConnectAPI/APIModels/RequestModels/TaskRequestModel.cs
@@ -17,7 +17,7 @@
         /// <summary>
         /// The progress
         /// </summary>
-        public decimal? Progress { get; set; }
+        public uint? Progress { get; set; }
 
         /// <summary>
         /// The status
